// src/models/DocumentDTO.ts
export interface DocumentDTO {
    accessCount: number;
    activeUserApprover: boolean;
    activeVersion: boolean;
    additionalComments: string;
    allowMuiltiCardsPerUser: boolean;
    approvalAndOr: boolean;
    approved: boolean;
    approvedDate: string;
    articleContent: string;
    atualizationId: number;
    backgroundColor: string;
    backgroundImage: string;
    bannerImage: string;
    cardDescription: string;
    colleagueId: string;
    colleagueName: string;
    companyId: number;
    convertDocumentType: number;
    crc: number;
    createDate: string;
    createDateInMilliseconds: number;
    datasetName: string;
    dateFormStarted: boolean;
    deleted: boolean;
    documentDescription: string;
    documentId: number;
    documentKeyWord: string;
    documentPropertyNumber: number;
    documentPropertyVersion: number;
    documentType: string;
    documentTypeId: string;
    downloadEnabled: boolean;
    draft: boolean;
    expirationDate: string;
    expiredForm: boolean;
    expires: boolean;
    externalDocumentId: string;
    favorite: boolean;
    fileURL: string;
    folderId: number;
    forAproval: boolean;
    hashAnnotations: string;
    iconId: number;
    iconPath: string;
    imutable: boolean;
    indexed: boolean;
    inheritApprovers: boolean;
    inheritSecurity: boolean;
    internalVisualizer: boolean;
    isEncrypted: boolean;
    keyWord: string;
    languageId: string;
    languageIndicator: string;
    lastModifiedDate: string;
    lastModifiedTime: string;
    metaListId: number;
    metaListRecordId: number;
    newStructure: boolean;
    notificationDays: number;
    onCheckout: boolean;
    parentDocumentId: number;
    pdfRenderEngine: string;
    permissionType: number;
    phisicalFile: string;
    phisicalFileSize: number;
    priority: number;
    privateColleagueId: string;
    privateDocument: boolean;
    protectedCopy: boolean;
    publicDocument: boolean;
    publisherId: string;
    publisherName: string;
    quota: number;
    relatedFiles: string;
    restrictionType: number;
    rowId: number;
    searchNumber: number;
    securityLevel: number;
    siteCode: string;
    socialDocument: string;
    tool: boolean;
    topicId: number;
    translated: boolean;
    UUID: string;
    updateIsoProperties: boolean;
    userAnswerForm: boolean;
    userNotify: boolean;
    userPermission: number;
    validationStartDate: string;
    version: number;
    versionDescription: string;
    versionOption: string;
    visualization: string;
    volumeId: string;
    watermarkId: number;
}